EBAC/Streamlit_Modulo_19_ Bank/requirements.txt
pandas==2.0.3
numpy==1.24.3
numpydoc==1.5.0
pandocfilters==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.2
python-slugify==5.0.2
python-snappy==0.6.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
streamlit==1.28.0
plotly==5.9.0
pillow==10.1.0